Allow for the job to automatically open in new tab
The motivation for this change is to allow for jobs to automatically
open in a new tab rather than having to directly right click and click
on "Open in a new tab". It is a more natural behavior to simply click
on each job element. Ctrl-click will also work by opening the tab but
keeping the current tab open.

By adding a target='_blank' the new link will open in a new tab.
However, when you simply click on it the parent div tag will still
receive the click and will redirect the current page. This behavior
can be removed by removing the click action from the parent element.
There is a little trick since the HTML is being changed the job body
will be inaccessible after the parent html is changed and vice versa.
So simply save the parent element and then change it afterwards.

Tested on Chrome and Firefox.